@@ -18,65 +18,39 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  animated_text_kit: ^2.2.0
-  cached_network_image: ^2.3.2+1
-  cupertino_icons: ^0.1.3
-  dartz: ^0.9.1
-  data_connection_checker: ^0.3.4
-  equatable: ^2.0.0
-  flutter_bloc: ^6.0.4
-  get_it: ^4.0.4
-  google_fonts: ^1.1.0
-  json_annotation: ^3.0.1
-  path: ^1.7.0
-  provider: ^4.3.2+2
-  retrofit: ^1.3.4
-  sqflite: ^1.3.1+1
+  animated_text_kit: ^4.2.1
+  cached_network_image: ^3.0.0
+  connectivity: ^3.0.6
+  cupertino_icons: ^1.0.3
+  dartz: ^0.10.0-nullsafety.2
+  equatable: ^2.0.2
+  flutter_bloc: ^7.0.1
+  get_it: ^7.1.3
+  google_fonts: ^2.1.0
+  json_annotation: ^4.0.1
+  path: ^1.8.0
+  provider: ^5.0.0
+  retrofit: ^2.0.0
+  sqflite: ^2.0.0+3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.10.0
-  json_serializable: ^3.4.1
-  retrofit_generator: ^1.3.7+6
+  build_runner: ^2.0.4
+  json_serializable: ^4.1.3
+  retrofit_generator: ^2.0.0
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-# The following section is specific to Flutter.
+dependency_overrides:
+  retrofit_generator:
+    git:
+      url: git@github.com:Chimerapps/retrofit.dart.git
+      ref: 9f90296751984b359937c38563da5b19db5550f5
+      path: generator
+
 flutter:
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - assets/images/movieguide.png
-  #   - assets/images/logo.svg
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
