@@ -1,7 +1,5 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
-import 'package:movieguide/data/repository/remote/api/api_config.dart';
 import 'package:movieguide/domain/entities/movie.dart';
 import 'package:movieguide/domain/entities/movie_kind.dart';
 import 'package:movieguide/domain/error/failure.dart';
@@ -12,7 +10,7 @@ class GetMovieUseCase
     extends BaseUseCase<Tuple2<List<Movie>, int>, GetMovieUseCaseParams> {
   final MovieRepository repository;
 
-  GetMovieUseCase({@required this.repository});
+  GetMovieUseCase({required this.repository});
 
   @override
   Future<Either<Failure, Tuple2<List<Movie>, int>>> call(
@@ -27,9 +25,9 @@ class GetMovieUseCaseParams extends Equatable {
   final MovieKind kind;
 
   GetMovieUseCaseParams({
-    this.apiKey = ApiConfig.API_KEY,
-    this.page = 1,
-    @required this.kind,
+    required this.apiKey,
+    required this.page,
+    required this.kind,
   });
 
   @override
