@@ -1,5 +1,4 @@
 import 'package:dartz/dartz.dart';
-import 'package:flutter/material.dart';
 import 'package:movieguide/domain/entities/movie.dart';
 import 'package:movieguide/domain/error/failure.dart';
 import 'package:movieguide/domain/repository/movie_repository.dart';
@@ -9,7 +8,7 @@ import 'base_usecase.dart';
 class LoadFavoriteMoviesUseCase extends BaseUseCase<List<Movie>, NoParams> {
   final MovieRepository repository;
 
-  LoadFavoriteMoviesUseCase({@required this.repository});
+  LoadFavoriteMoviesUseCase({required this.repository});
 
   @override
   Future<Either<Failure, List<Movie>>> call(NoParams params) async {
