@@ -1,6 +1,5 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 import 'package:movieguide/domain/error/failure.dart';
 import 'package:movieguide/domain/repository/movie_repository.dart';
 import 'package:movieguide/domain/usecases/base_usecase.dart';
@@ -9,7 +8,7 @@ class RemoveMovieFromFavoriteUseCase
     extends BaseUseCase<bool, RemoveMovieFromFavoriteParams> {
   final MovieRepository repository;
 
-  RemoveMovieFromFavoriteUseCase({@required this.repository});
+  RemoveMovieFromFavoriteUseCase({required this.repository});
 
   @override
   Future<Either<Failure, bool>> call(RemoveMovieFromFavoriteParams params) {
@@ -20,7 +19,7 @@ class RemoveMovieFromFavoriteUseCase
 class RemoveMovieFromFavoriteParams extends Equatable {
   final String movieId;
 
-  RemoveMovieFromFavoriteParams({@required this.movieId});
+  RemoveMovieFromFavoriteParams({required this.movieId});
   @override
   List<Object> get props => [movieId];
 }
