@@ -1,6 +1,5 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:movieguide/di.dart';
 import 'package:movieguide/domain/entities/movie_kind.dart';
@@ -9,7 +8,7 @@ import 'package:movieguide/presentation/home/views/subviews/bottom_loading.dart'
 import 'package:movieguide/presentation/home/views/subviews/movie_item.dart';
 
 class TabNewest extends StatelessWidget {
-  const TabNewest({Key key}) : super(key: key);
+  const TabNewest({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -28,7 +27,7 @@ class NewestMovies extends StatefulWidget {
 
 class _NewestMoviesState extends State<NewestMovies> {
   final ScrollController _scrollController = ScrollController();
-  MovielistBloc _movielistBloc;
+  MovielistBloc? _movielistBloc;
   @override
   void initState() {
     super.initState();
@@ -40,7 +39,8 @@ class _NewestMoviesState extends State<NewestMovies> {
     double maxScroll = _scrollController.position.maxScrollExtent;
     double currentScroll = _scrollController.position.pixels;
     double delta = 200.0;
-    if (currentScroll >= maxScroll - delta && !_movielistBloc.isLoading) {
+    if (currentScroll >= maxScroll - delta &&
+        !(_movielistBloc?.isLoading ?? false)) {
       _movielistBloc?.fectchMovies(MovieKind.newest);
     }
   }
@@ -49,7 +49,7 @@ class _NewestMoviesState extends State<NewestMovies> {
   void dispose() {
     super.dispose();
     _scrollController.dispose();
-    _movielistBloc.close();
+    _movielistBloc?.close();
   }
 
   @override
@@ -59,35 +59,36 @@ class _NewestMoviesState extends State<NewestMovies> {
         if (state is MovielistInitial) {
           return Center(child: CircularProgressIndicator());
         } else if (state is MovielistError) {
-          SchedulerBinding.instance.addPostFrameCallback((timeStamp) {
-            showDialog(
-              context: context,
-              builder: (_) {
-                return AlertDialog(
-                    title: Text("Something wrongs !"),
-                    content: Text(state.errorMessage),
-                    actions: [
-                      FlatButton(
-                        child: Text("OK"),
-                        onPressed: () {
-                          Navigator.of(context).pop();
-                        },
-                      ),
-                    ]);
-              },
-            );
-          });
+          // SchedulerBinding.instance.addPostFrameCallback((timeStamp) {
+          //   showDialog(
+          //     context: context,
+          //     builder: (_) {
+          //       return AlertDialog(
+          //           title: Text("Something wrongs !"),
+          //           content: Text(state.errorMessage),
+          //           actions: [
+          //             FlatButton(
+          //               child: Text("OK"),
+          //               onPressed: () {
+          //                 Navigator.of(context).pop();
+          //               },
+          //             ),
+          //           ]);
+          //     },
+          //   );
+          // });
           return Container();
         }
-        return _buildMovieList(context, state);
+        return _buildMovieList(context, state as MovielistLoaded);
       },
     );
   }
 
   Widget _buildMovieList(BuildContext context, MovielistLoaded state) {
     final movies = state.movies;
+    if (_movielistBloc == null) return Container();
     return RefreshIndicator(
-      onRefresh: () => _movielistBloc.refresh(MovieKind.newest),
+      onRefresh: () => _movielistBloc!.refresh(MovieKind.newest),
       child: CustomScrollView(
         controller: _scrollController,
         slivers: [
