@@ -1,5 +1,5 @@
 class ApiConfig {
   static const String BASE_URL = "https://api.themoviedb.org/3";
   static const String IMAGE_BASE_URL = "http://image.tmdb.org/t/p/w342";
-  static const String API_KEY = "";
+  static const String API_KEY = "0cff4011ebb83cb7a698ffcacfc0a336";
 }
