@@ -1,7 +1,8 @@
-import 'package:data_connection_checker/data_connection_checker.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:get_it/get_it.dart';
 import 'package:movieguide/core/platform/network_info.dart';
 import 'package:movieguide/data/model/mapper/movie_data_mapper.dart';
+import 'package:movieguide/data/model/mapper/movie_detail_mapper.dart';
 import 'package:movieguide/data/repository/local/api/db/app_database.dart';
 import 'package:movieguide/data/repository/local/api/db/impl/app_database_impl.dart';
 import 'package:movieguide/data/repository/local/movie_local_datasource.dart';
@@ -11,6 +12,7 @@ import 'package:movieguide/data/repository/remote/api/movie_api.dart';
 import 'package:movieguide/data/repository/remote/api/service_generator.dart';
 import 'package:movieguide/data/repository/remote/movie_remote_datasource.dart';
 import 'package:movieguide/domain/usecases/add_movie_to_favorite_usecase.dart';
+import 'package:movieguide/domain/usecases/get_movie_detail_usecase.dart';
 import 'package:movieguide/domain/usecases/get_movies_usecase.dart';
 import 'package:movieguide/domain/usecases/load_favorite_movie_usecase.dart';
 import 'package:movieguide/domain/usecases/remove_movie_from_favorite.dart';
@@ -22,14 +24,15 @@ final getIt = GetIt.instance;
 Future<void> setupDi() async {
   _coreModule();
   _networkModule();
+  _mapperModule();
   _repositoryModule();
   _usecaseModule();
 }
 
 _coreModule() {
-  getIt.registerSingleton(DataConnectionChecker());
+  getIt.registerSingleton(Connectivity());
   getIt.registerSingleton<NetworkInfo>(
-    NetworkInfoImpl(getIt.get<DataConnectionChecker>()),
+    NetworkInfoImpl(getIt.get<Connectivity>()),
   );
 }
 
@@ -38,8 +41,12 @@ _networkModule() {
   getIt.registerSingleton(MovieApi(getIt.get()));
 }
 
-_repositoryModule() {
+_mapperModule() {
   getIt.registerSingleton(MovieDataMaper());
+  getIt.registerSingleton(MovieDetailMapper());
+}
+
+_repositoryModule() {
   getIt.registerSingleton<AppDatabase>(AppDataBaseImpl.instance);
   getIt.registerSingleton(MovieRemoteDataSource(movieApi: getIt.get()));
   getIt.registerSingleton(MovieLocalDataSource(getIt.get()));
@@ -48,6 +55,7 @@ _repositoryModule() {
     localDataSource: getIt.get(),
     networkInfo: getIt.get(),
     movieDataMapper: getIt.get(),
+    movieDetailMapper: getIt.get(),
   ));
 }
 
@@ -62,4 +70,5 @@ _usecaseModule() {
   getIt.registerFactory(
     () => LoadFavoriteMoviesUseCase(repository: getIt.get()),
   );
+  getIt.registerFactory(() => GetMovieDetailUseCase(repository: getIt.get()));
 }
