@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 
-class BaseModel with ChangeNotifier {
+enum ViewState { Busy, Idle }
+
+abstract class BaseModel extends ChangeNotifier {
   ViewState _state = ViewState.Idle;
   ViewState get state => _state;
-
   void setViewState(ViewState newState) {
     _state = newState;
-    notifyListeners();
   }
 }
-
-enum ViewState { Idle, Busy }
