@@ -1,9 +1,6 @@
-import 'package:movieguide/data/repository/remote/api/api_config.dart';
 import 'package:movieguide/domain/entities/movie.dart';
-import 'package:movieguide/domain/entities/movie_kind.dart';
 import 'package:movieguide/domain/usecases/add_movie_to_favorite_usecase.dart';
 import 'package:movieguide/domain/usecases/base_usecase.dart';
-import 'package:movieguide/domain/usecases/get_movies_usecase.dart';
 import 'package:movieguide/domain/usecases/load_favorite_movie_usecase.dart';
 import 'package:movieguide/domain/usecases/remove_movie_from_favorite.dart';
 import 'package:movieguide/presentation/base/base_model.dart';
@@ -13,9 +10,9 @@ class FavoriteMovieChangeNotifier extends BaseModel {
   final RemoveMovieFromFavoriteUseCase removeMovieFromFavoriteUseCase;
   final LoadFavoriteMoviesUseCase favoriteMoviesUseCase;
   FavoriteMovieChangeNotifier({
-    this.addMovieToFavoriteUseCase,
-    this.removeMovieFromFavoriteUseCase,
-    this.favoriteMoviesUseCase,
+    required this.addMovieToFavoriteUseCase,
+    required this.removeMovieFromFavoriteUseCase,
+    required this.favoriteMoviesUseCase,
   });
 
   final List<Movie> _favoriteMovies = [];
